client/pyproject.toml
[tool.poetry]
name = "client"
version = "0.1.0"
description = ""
authors = ["xkumiyu <xkumiyu@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.8"
grpcio = "^1.34.0"
tensorflow = "^2.3.1"
tensorflow-serving-api = "^2.3.0"
opencv-python = "^4.4.0"
tqdm = "^4.54.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
black = "^20.8b1"
isort = "^5.6.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
